// apps/backend/src/platforms/mobile/schema.ts
export type MobileLocator = {
  ref?: string;
  accessibilityId?: string;
  text?: string;
  name?: string;
};